Comments from code review resolved, rest of creds moved to properties file.
