+++ src/test/java/com/softserve/edu/greencity/ui/tools/GMailLogin.java
@@ -10,6 +10,9 @@ import org.slf4j.LoggerFactory;
 
 import java.awt.*;
 import java.awt.datatransfer.DataFlavor;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.util.Properties;
 import java.util.concurrent.TimeUnit;
 
 public class GMailLogin {
@@ -21,7 +24,7 @@ public class GMailLogin {
     private WebElement loginPassword;
     private WebElement signInButton;
 
-
+    private Properties property = new Properties();
 
     public final static String URL = "https://accounts.google.com/signin/v2/identifier?" +
             "continue=https%3A%2F%2Fmail.google.com%2Fmail%2F&service=mail&sacu=1&rip=1&" +
@@ -32,12 +35,19 @@ public class GMailLogin {
 
         this.driver = driver;
         driver.get(URL);
+
+        try {
+            final FileInputStream fis = new FileInputStream("src/test/resources/credentials.properties");
+            property.load(fis);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
     }
 
 
    public GMailLogin enterEmail(){
         loginEmail = driver.findElement(By.id("identifierId"));
-        loginEmail.sendKeys("GCSignUpUser@gmail.com");
+        loginEmail.sendKeys(property.getProperty("emailForRegistration"));
         return this;
    }
 
@@ -49,7 +59,7 @@ public class GMailLogin {
 
     public GMailLogin enterPassword(){
         loginPassword = driver.findElement(By.name("password"));
-        loginPassword.sendKeys("Error911");
+        loginPassword.sendKeys(property.getProperty("passwordToGmailBox"));
         return this;
     }
 
@@ -68,6 +78,15 @@ public class GMailLogin {
             e.printStackTrace();
         }
         return this;
+    }
+
+    public GMailBox logInGMail()  {
+        GMailLogin loginPage = new GMailLogin(driver);
+        loginPage.enterEmail()
+                .clickNext()
+                .enterPassword()
+                .clickSignInButton();
+        return new GMailBox(driver);
     }
 
 }
